changed StandardTier into component. removed from router. incerted component to Subscription page.
